// src/client/eco-params.ts
export interface EcoParams {
    url: string
    consumerKey: string
    consumerSecret: string
    accessToken: string
    tokenSecret: string

    dbId: string
    loginUsuarioEco: boolean | true
    empresaEco: string
    usuarioEco: string
    senhaEco: string
    idLocal: string
    codigoValidacaoLocal: string

}